Time Contextual Headers [1/2]
This commit adds image headers to the notification panel header,
based on the date/time.

There is a default provider that provides sky pictures
as well as special events like Christmas and new
year's eve.

More will be coming, with eventually a way to use an external
(as in, Internet) provider for those pictures.

Signed-off-by: NoelMacwan <yulespencer@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarHeaderView.java