Fix FragmentManager to save state of a Fragment only when needed
FragmentManager to save state of a Fragment only when it has never been
saved or if it an active fragment with a view. This ensures that
onSaveInstanceState isn't invoked (for the second time) on a Fragment
in back stack without a view. Saving the instance of framgents multiple
times (caused by multiple screen orientation changes) can break some
applications as each orientation causes a new instance of the fragment
to be created and its member data structures would be null.

Change-Id: I202536433b4f5e3cd5e2bc7d82e75a3df8c6952e
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>